requirements.txt
google-api-core==1.21.0
google-api-python-client==1.10.0
google-auth==1.19.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
PyDrive==1.3.1
python-dotenv==0.13.0
requests==2.23.0
vk-api==11.8.0
python-telegram-bot==12.7
urlextract==1.0.0




